refactor: 파일구조를 변경하였습니다.
