+++ Box42/Toolbar/View/DisplayURL/DisplayURLInToolbar.swift
@@ -12,6 +12,7 @@ import SnapKit
 class DisplayURLInToolbar: NSView {
     var URLTextfield: DisplayURLTextfield = DisplayURLTextfield()
     var originalString: String = ""
+    var icon = MenubarViewController()
     
     override init(frame frameRect: NSRect) {
         super.init(frame: .zero)
@@ -87,13 +88,7 @@ extension DisplayURLInToolbar: WKNavigationDelegate {
             if showURLString.count > 1 {
                 URLTextfield.stringValue = (showURLString[1] ?? "")
             }
+            print(originalString)
         }
-    }
-}
-
-extension DisplayURLInToolbar {
-    func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-        print("Navigation finished")
-        updateURL()
     }
 }
